SpikeAnalysis

Documentation for SpikeAnalysis.

SpikeAnalysis.RasterTrigSylls

SpikeAnalysis.TrigSet

SpikeAnalysis.WarpPlan

SpikeAnalysis.add_pres

SpikeAnalysis.align_trigs

SpikeAnalysis.align_trigs

SpikeAnalysis.call_bouts

SpikeAnalysis.clip_trigs

SpikeAnalysis.spike_clip_fixed

SpikeAnalysis.trig_data

SpikeAnalysis.xcorr_discrete_normed

SpikeAnalysis.RasterTrigSylls — Type

RasterTrigSylls{D, M<:MarkedInterval{D, <:Any}}

Represents a "trigger", which is a sequence of syllables in a longer sequence that matches some pattern. This is akin to the trigger of an oscilloscope, or electro_gui.

SpikeAnalysis.TrigSet — Type

TrigSet{D, R<:RasterTrigSylls{D, <:Any}

Represents a group of triggers, with some time around each trigger.

See Also: RasterTrigSylls

SpikeAnalysis.WarpPlan — Type

WarpPlan{R, S<:Union{Nothing, Float64}}

Type for storing a piece of a piece-wise linear transform (warping) of points.

SpikeAnalysis.add_pres — Method

function add_pres(
    rec::NakedInterval{E},
    sylls::AbstractVector{M},
    trigsylls::AbstractVector{<:String};
    pre_syll::Real = 0.03,
) where {E, M<:MarkedInterval{E, <:Any}} -> expanded, points

Expand syllables, syll with label in trigsylls, by pre_syll, up to the last syll or the bound_start.

Requires input to be sorted and not overlapping.

Returns a tuple where the first elements, expanded, is an array of interval sets with the original syllable joined with the "pre" period. The second element, points, is an array of NakedPoints with points for syllable onset and offset defined on an expanded interval, including the "pre" period.

SpikeAnalysis.align_trigs — Method

align_trigs(raw::RasterTrigSylls)

Takes a RasterTrigSylls, and aligns its syllable times to the onset of the first triggered syllable. Returns a new, aligned, RasterTrigSylls.

See also: align_trigs(::TrigSet)

SpikeAnalysis.align_trigs — Method

align_trigs(raw::TrigSet)

Takes a TrigSet, and aligns each RasterTrigSylls therein to the onset time of its first triggered syllable.

See also: align_trigs(::RasterTrigSylls)

SpikeAnalysis.call_bouts — Function

Finds bouts of song

SpikeAnalysis.clip_trigs — Method

clip_trigs(raw::TrigSet)

Clips syllables in a TrigSet to its expanded interval.

SpikeAnalysis.spike_clip_fixed — Function

Get subsection of signal centered at spike_idx, fill empty spots with NaN

SpikeAnalysis.trig_data — Function

trig_data(
    trig_seq::AbstractVector{<:AbstractString},
    labels_syl::AbstractVector{<:AbstractString},
    offs_syl::AbstractVector{<:Real},
    durs_syl::AbstractVector{<:Real},
    syls::AbstractVector{<:MarkedInterval},
    pre::Real,
    post_expand::Union{Real, Nothing} = nothing;
    post::Union{Nothing, Real} = nothing
)

Finds occurrences of the sub-sequence specified by trig_seq in the sequence labels_syl. Returns a TrigSet. post specifies time to stop after trig onset, while post_expand specifies time stop after trig offset.

SpikeAnalysis.xcorr_discrete_normed — Method

xcorr_discrete_normed(us, vs, durs; binsize=0.005, maxdiff=nothing, edgecorrect=true, closed=:left)

Calculate the normalized cross-correlation of two sets of discrete time points.

Parameters

us::AbstractVector{<:AbstractVector{<:Number}}: A collection of sets of discrete time points.

vs::AbstractVector{<:AbstractVector{<:Number}}: A collection of sets of discrete time points corresponding to us.

durs::AbstractVector{<:Number}: Duration of each subsection. Must have the same length as us and vs.

binsize::Number=0.005: Bin size for the histogramming of time differences.

maxdiff::Union{Number, Nothing}=nothing: Maximum time difference to consider. If set to nothing, it is determined by the maximum of durs.

edgecorrect::Bool=true: Whether to apply edge correction to the histogram.

closed=:left: Which side of the bin intervals is closed. Either :left or :right.

Returns

counts: Normalized cross-correlation histogram counts.

centers: Centers of the bins in the histogram.

Notes

The inputs us and vs must be sorted and unique. If the points are not unique, the normalization will be incorrect.

Throws an error if maxdiff is provided and is less than binsize.

Throws an ArgumentError if the lengths of us, vs, and durs are not the same.